requirements.txt
pytest
python-dotenv
requests
requests_mock